Rename test files for clippy
